[Refactor] #935 Renamed player-classes-types.h to player-class-types.h
