+++ src/birth/birth-explanations-table.h
@@ -1,6 +1,6 @@
 ﻿#pragma once
 
-#include "player/player-classes-types.h"
+#include "player/player-class-types.h"
 #include "player/player-personality.h"
 #include "player/player-race-types.h"
 #include "realm/realm-names-table.h"
