+++ src/birth/initial-equipments-table.h
@@ -1,6 +1,6 @@
 ﻿#pragma once
 
-#include "player/player-classes-types.h"
+#include "player/player-class-types.h"
 #include "system/angband.h"
 
 extern byte player_init[MAX_CLASS][3][2];
