+++ src/birth/quick-start.h
@@ -1,6 +1,6 @@
 ﻿#pragma once
 
-#include "player/player-classes-types.h"
+#include "player/player-class-types.h"
 #include "player/player-personality-types.h"
 #include "player/player-race-types.h"
 #include "system/angband.h"
