+++ src/cmd-io/cmd-menu-content-table.cpp
@@ -1,5 +1,5 @@
 ﻿#include "cmd-io/cmd-menu-content-table.h"
-#include "player/player-classes-types.h"
+#include "player/player-class-types.h"
 #include "util/int-char-converter.h"
 
 special_menu_content special_menu_info[MAX_SPECIAL_MENU_NUM] = {
