+++ src/player/player-race.h
@@ -6,7 +6,7 @@
 
 #include "object-enchant/tr-types.h"
 #include "player-ability/player-abillity-types.h"
-#include "player/player-classes-types.h"
+#include "player/player-class-types.h"
 #include "player/player-race-types.h"
 #include "system/angband.h"
