+++ src/system/building-type-definition.h
@@ -2,7 +2,7 @@
 
 /* 人畜無害なenumヘッダを先に読み込む */
 #include "object/tval-types.h"
-#include "player/player-classes-types.h"
+#include "player/player-class-types.h"
 #include "player/player-race-types.h"
 #include "realm/realm-types.h"
 #include "system/angband.h"
