+++ src/system/player-type-definition.h
@@ -2,7 +2,7 @@
 
 #include "player-ability/player-abillity-types.h"
 #include "player/player-race-types.h"
-#include "player/player-classes-types.h"
+#include "player/player-class-types.h"
 #include "player/player-personality-types.h"
 #include "system/angband.h"
 #include "system/system-variables.h"
